Feature: "`linear_x`" ---------------------
Since the X axis is often a bunch of date strings (or maybe date objects), we may want labels that are some human-readable text representation of a date (instead of a huge number, like unix epoch).

We take advantage of the fact that in these cases, the X values are evenly spaced along the axis. We just calculate the distance between them (which is constant) and pick labels out of the X values supplied "every N" values.